% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffmetadata-package.R
\docType{package}
\name{ffmetadata}
\alias{ffmetadata}
\alias{ffmetadata-package}
\title{ffmetadata}
\description{
This package provides functions to access to metadata for data from the
Fragile Families Project.
}
\section{Search functions}{

To search or filter variables by string:
\itemize{
  \item{\code{\link{search_metadata}}}
  \item{\code{\link{filter_metadata}}}
}
}

\section{Information functions}{

To get more detail on a variable:
\itemize{
  \item{\code{\link{select_metadata}}}
}
}

