% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{filter_metadata}
\alias{filter_metadata}
\title{Filter Metadata}
\usage{
filter_metadata(filter_list = list(), ...)
}
\arguments{
\item{filter_list}{a named list of variables to filter metadata on. See
details for valid field names.}

\item{...}{additional value combinations to filter on. See details for
valid field names.}
}
\value{
returns list of all variable names that match a set of given
filter values
}
\description{
filter_metadata allows users to retrieve a list of variable
names based on a set of filter values for variable categories
}
\details{
List of valid field names:
\itemize{
    \item{data_source}
    \item{data_type}
    \item{group_id}
    \item{group_subid}
    \item{id}
    \item{label}
    \item{leaf}
    \item{name}
    \item{old_name}
    \item{respondent}
    \item{responses}
    \item{scope}
    \item{section}
    \item{topic}
    \item{umbrella}
    \item{warning}
    \item{wave}
}
}
\examples{
filter_test <- filter_metadata(wave = 3, data_source = "constructed", data_type = "bin")
}
