% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{filterMetadata}
\alias{filterMetadata}
\title{filterMetadata}
\usage{
filterMetadata(...)
}
\arguments{
\item{...}{}
}
\value{
returns data frame where each observation is a
variable
}
\description{
Modified from PHP Web API definition: filterMetadata returns a data frame
of variables given a set of filter values for variable categories,
or the entire set of metadata as a data frame if no filters are provided
}
\examples{
filter_test <- filterMetadata(wave = 3, source = "constructed", type = "bin")
}
