% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{search_metadata}
\alias{search_metadata}
\title{Search Metadata}
\usage{
search_metadata(query, field_name)
}
\arguments{
\item{query}{a substring searched for in the given variable field}

\item{field_name}{the field in which the query is searched for.
See details for valid field names.}
}
\value{
returns list of names of all variables that have a substring-matched value within
a given field
}
\description{
search_metadata allows users to retrieve a list of
variable names based on whether or not those variables contain a given query
within a given field
}
\details{
List of valid field names:

List of valid field names:
\itemize{
    \item{data_source}
    \item{data_type}
    \item{group_id}
    \item{group_subid}
    \item{id}
    \item{label}
    \item{leaf}
    \item{name}
    \item{old_name}
    \item{respondent}
    \item{responses}
    \item{scope}
    \item{section}
    \item{topic}
    \item{umbrella}
    \item{warning}
    \item{wave}
}
}
\examples{
search_test <- search_metadata(query = "oc", field_name = "data_type")
}
