% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchMetadata.R
\name{searchMetadata}
\alias{searchMetadata}
\title{searchMetadata}
\usage{
searchMetadata(query, fieldName)
}
\arguments{
\item{query}{a substring searched for in the given search category}

\item{fieldName}{the search category}
}
\value{
returns data frame where each observation is a variable
}
\description{
Modified from PHP Web API: searchMetadata returns a
data frame of variables given a query string and a search category,
or the entire set of metadata as a data frame if query is empty
}
\examples{
searchTest <- searchMetadata("policing", "topic1")
}
