% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{select_metadata}
\alias{select_metadata}
\title{Select Metadata}
\usage{
select_metadata(variable_name = NULL, fields = NULL)
}
\arguments{
\item{variable_name}{name of variable}

\item{field_name}{specific field of variable to be accessed. See
details for valid field names}
}
\value{
returns string with value of a given field if field is specified,
returns all metadata for variable as a named list if field is unspecified
}
\description{
select_metadata allows users to retrieve the value of
a specific field of a given variable, or
the entire variable metadata if no field is specified
}
\details{
List of valid field names:
\itemize{
    \item{data_source}
    \item{data_type}
    \item{group_id}
    \item{group_subid}
    \item{id}
    \item{label}
    \item{leaf}
    \item{name}
    \item{old_name}
    \item{respondent}
    \item{responses}
    \item{scope}
    \item{section}
    \item{topic}
    \item{umbrella}
    \item{warning}
    \item{wave}
}
}
\examples{
select1 <- select_metadata(variable_name = "ce3agefc")
select2 <- select_metadata(variable_name = "ce3agefc", field_name = "data_type")
}
