% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectMetadata.R
\name{selectMetadata}
\alias{selectMetadata}
\title{selectMetadata}
\usage{
selectMetadata(varName, fieldName)
}
\arguments{
\item{varName}{# name of variable}

\item{fieldName}{# specific field of variable}
}
\value{
returns string with value if field is specified,
returns list of fields for variable if field is unspecified
}
\description{
Modified from PHP Web API definition: selectMetadata
returns a specific field of a given variable, or
the entire variable if no field is specified
}
\examples{
selectNoFieldName <- selectMetadata("ce3agefc")
selectWithFieldName <- selectMetadata("ce3agefc", "type")
}
