% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{filter_metadata}
\alias{filter_metadata}
\title{Filter Metadata}
\usage{
filter_metadata(filter_list = list(), ...)
}
\arguments{
\item{...}{}
}
\value{
returns data frame where each observation is a
variable
}
\description{
Modified from PHP Web API definition: filterMetadata returns a data frame
of variables given a set of filter values for variable categories,
or the entire set of metadata as a data frame if no filters are provided
}
\examples{
filter_test <- filter_metadata(wave = 3, source = "constructed", type = "bin")
}

