% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{search_metadata}
\alias{search_metadata}
\title{Search Metadata}
\usage{
search_metadata(query = NULL, field_name = NULL)
}
\arguments{
\item{query}{a substring searched for in the given search category}

\item{field_name}{the search category}
}
\value{
returns data frame where each observation is a variable
}
\description{
Modified from PHP Web API: searchMetadata returns a
data frame of variables given a query string and a search category,
or the entire set of metadata as a data frame if query is empty
}
\examples{
search_test <- search_metadata(query = "policing", field_name = "topic1")
}

