% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{select_metadata}
\alias{select_metadata}
\title{Select Metadata}
\usage{
select_metadata(variable_name = NULL, field_name = NULL)
}
\arguments{
\item{variable_name}{name of variable}

\item{field_name}{specific field of variable}
}
\value{
returns string with value if field is specified,
returns list of fields for variable if field is unspecified
}
\description{
Modified from PHP Web API definition: select_metadata
returns a specific field of a given variable, or
the entire variable if no field is specified
}
\examples{
select1 <- select_metadata(variable_name = "ce3agefc")
select2 <- select_metadata(variable_name = "ce3agefc", field_name = "type")
}
